### What I did

before:

```
CPU times: user 1.85 s, sys: 184 ms, total: 2.03 s
Wall time: 2.05 s
```

after:

```
In [1]: %time import ape_polygon
CPU times: user 7.99 ms, sys: 3.13 ms, total: 11.1 ms
Wall time: 10.1 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)